backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
transformers==4.36.0
torch==2.1.0
torchvision==0.16.0
accelerate==0.25.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
pydantic==2.5.0
